perf(webp): trellis_quantize_block i16 in/out (stage B, m6 lever)
trellis_quantize_block's coefficient block is now i16 in/out like
libwebp's TrellisQuantizeBlock (int16_t in[16]/out[16]); the DP widens
per-coefficient reads into i32 locals and narrows the level*q
reconstruction store. This deletes every caller-side widen shim the
previous slice introduced (i4_trellis_quantize_recon, the I16 final
pass, i16_rd_trellis_quantize, the recorder's verify + requant arms) —
those callers now hand their i16 blocks to trellis directly, and the
trellis-arm reconstruction reuses the dequantized values it writes back
(idct4x4_i16) instead of re-deriving them.

Also swept the I4 final pass (transform_luma_blocks_4x4) to a single
raw-DCT narrow feeding coeffs store + quantizer + trellis + the i16
fused IDCT-add, and deleted the now-unused i32 idct_add_residue_inplace
dispatch family (the decoder keeps its _sse2_inner/_neon_inner paths).

Byte gate: output_hash COMBINED 958f376a6c8b118f unchanged. Suites:
x86-64 741 passed, i686 715 passed, wasm32 builds, clippy x3 clean.

+++ src/encoder/cost/mod.rs
@@ -322,7 +322,7 @@ mod tests {
 
         // Test case: DC dominant block (natural order)
         let block = [
-            500i32, 50, 25, 10, // row 0
+            500i16, 50, 25, 10, // row 0
             5, 3, 2, 1, // row 1
             0, 0, 0, 0, // row 2
             0, 0, 0, 0, // row 3
@@ -354,7 +354,7 @@ mod tests {
         let mut simple_out = [0i32; 16];
         for i in 0..16 {
             let j = VP8_ZIGZAG[i]; // Convert zigzag position to natural
-            simple_out[i] = matrix.quantize_coeff(block[j], j);
+            simple_out[i] = matrix.quantize_coeff(i32::from(block[j]), j);
         }
 
         // Both should produce non-zero DC coefficient
